Перенес отрисовку экстра карточек в метод render класса pageController
